+++ client/src/App.jsx
@@ -15,7 +15,7 @@ import Registration from "./components/login-reg-components/Registration";
 import Profile from "./components/Profile";
 import Grid from "@mui/material/Grid";
 import Box from "@mui/material/Box";
-import { UserContextProvider, UserContext } from "./components/UserContext";
+import { AuthContextProvider, AuthContext } from "./components/AuthContext";
 import { Redirect } from "react-router-dom";
 
 function App() {
@@ -39,44 +39,46 @@ function App() {
 
   return (
     <ThemeProvider theme={siteTheme}>
-      <Box className="App" sx={{ bgcolor: "primary.light" }}>
-        <BrowserRouter>
-          <Switch>
+      <AuthContextProvider>
+        <Box className="App" sx={{ bgcolor: "primary.light" }}>
+          <BrowserRouter>
+            {/* <Switch> */}
             <Route exact path="/" component={Landing} />
             <Route exact path="/login" component={Login} />
             <Route exact path="/register" component={Registration} />
-          </Switch>
-          <UserContextProvider>
+            {/* </Switch> */}
             {/* <UserContext.Consumer> */}
-            <Route exact path="/dashboard" component={Dashboard} />
+            <Switch>
+              <Route exact path="/dashboard" component={Dashboard} />
 
-            <Redirect to="/" />
-            {/* </UserContext.Consumer> */}
-            <Route exact path="/posts/create">
-              <RegisteredNavBar />
-              <Grid container spacing={3} m={1} p={1}>
-                <Grid item xs={12} sm={6} md={4}>
-                  <Profile />
+              <Redirect to="/" />
+              {/* </UserContext.Consumer> */}
+              <Route exact path="/posts/create">
+                <RegisteredNavBar />
+                <Grid container spacing={3} m={1} p={1}>
+                  <Grid item xs={12} sm={6} md={4}>
+                    <Profile />
+                  </Grid>
+                  <Grid item xs={12} sm={6} md={8}>
+                    <CreatePost />
+                  </Grid>
                 </Grid>
-                <Grid item xs={12} sm={6} md={8}>
-                  <CreatePost />
+              </Route>
+              <Route exact path="/posts/edit/:id">
+                <RegisteredNavBar />
+                <Grid container spacing={3} m={1} p={1}>
+                  <Grid item xs={12} sm={6} md={4}>
+                    <Profile></Profile>
+                  </Grid>
+                  <Grid item xs={12} sm={6} md={8}>
+                    <EditPost active={true} />
+                  </Grid>
                 </Grid>
-              </Grid>
-            </Route>
-            <Route exact path="/posts/edit/:id">
-              <RegisteredNavBar />
-              <Grid container spacing={3} m={1} p={1}>
-                <Grid item xs={12} sm={6} md={4}>
-                  <Profile></Profile>
-                </Grid>
-                <Grid item xs={12} sm={6} md={8}>
-                  <EditPost active={true} />
-                </Grid>
-              </Grid>
-            </Route>
-          </UserContextProvider>
-        </BrowserRouter>
-      </Box>
+              </Route>
+            </Switch>
+          </BrowserRouter>
+        </Box>
+      </AuthContextProvider>
     </ThemeProvider>
   );
 }
